Move settimer to new folder
