+++ src/app/pages/SetTimer/SetTimer.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Slider from './Slider';
+import Slider from './SetTimer';
 
 export default {
   title: 'Component/Slider',
